+++ CHANGELOG.rst
@@ -1,7 +1,9 @@
 1.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Provide as async utility as it allows us to close connections when the object
+  is destroyed
+  [vangheem]
 
 
 1.0.0 (2017-04-24)